Module openvr::compositor

The Compositor implements distortion, prediction, synchronization and other subtle issues that can be a challenge to get operating properly for a solid VR experience.

Applications call WaitGetPoses to get the set of poses used to render the camera and other tracked objects, render the left and right eyes as normal (using the info provided by System) and finally submit those undistorted textures for the Compositor to display on the output device.

It is recommended that you continue presenting your application's own window, reusing either the left or right eye camera render target to draw a single quad (perhaps cropped to a lower fov to hide the hidden area mask).

Re-exports

pub use self::texture::Texture;

Modules

compositor_error

texture

Structs

CompositorError

WaitPoses