Struct openvr::render_models::IVRRenderModels

pub struct IVRRenderModels(pub *const ());

Methods

impl IVRRenderModels

unsafe fn from_raw(ptr: *const ()) -> Self

fn get_count(&self) -> u32

Returns the amount of render models available

fn get_name(&self, index: u32) -> String

Returns the name of an available render model

fn load(&self, name: String) -> Result<RenderModel, Error<EVRRenderModelError>>

Loads an render model into local memory blocks the thread and waits until driver responds with model

fn load_async(&self, name: String) -> Result<RenderModel, Error<EVRRenderModelError>>

Loads an render model into local memory When called for the first time openvr will start to load the model into memory In the mean time this call will respond with EVRRenderModelError_VRRenderModelError_Loading It is designed to be used wihtin the render loop as it won't block the user, for sync usage use load()