Struct openvr::Context

pub struct Context { /* fields omitted */ }

Entry points into OpenVR.

At most one of this object may exist at a time.

See safety notes in init.

Methods

impl Context

pub fn system(&self) -> Result<System, InitError>

pub fn compositor(&self) -> Result<Compositor, InitError>

pub fn render_models(&self) -> Result<RenderModels, InitError>

impl Context

pub unsafe fn shutdown(&self)

Shut down OpenVR. Repeated calls are safe.

Called implicitly by Context::drop.

Safety

This must be called before shutting down the graphics API, or OpenVR may invoke undefined behavior by attempting to free graphics resources.

No calls to other OpenVR methods may be made after this has been called unless a new Context is first constructed.