Module openvr::system

The System interface provides access to display configuration information, tracking data, controller state, events, and device properties. It is the main interface of OpenVR.

Re-exports

pub use self::event::Event;

pub use self::event::EventInfo;

Modules

event

tracked_property_error

Structs

DistortionCoordinates

HiddenAreaMesh
A triangle mesh containing geometry determined by HiddenAreaMeshType.

RawProjection
Values represent the tangents of the half-angles from the center view axis

TrackedPropertyError

Enums

HiddenAreaMeshType