Module openvr::system::event

Structs

Chaperone

Controller
Controller button events

EventInfo

Ipd

Keyboard

Mouse
Simulated mouse events in overlay space

Notification
notification related events. Details will still change at this point

Overlay

Process

Property

Scroll
Simulated mouse wheel scroll in overlay space

Status

TouchPadMove
When in mouse input mode you can receive data from the touchpad, these events are only sent if the user's finger is on the touchpad (or just released from it)

Enums

Event

Type Definitions

PropertyContainerHandle