Enum openvr::system::event::Event

pub enum Event {
    TrackedDeviceActivated,
    TrackedDeviceDeactivated,
    TrackedDeviceUpdated,
    TrackedDeviceUserInteractionStarted,
    TrackedDeviceUserInteractionEnded,
    IpdChanged,
    EnterStandbyMode,
    LeaveStandbyMode,
    TrackedDeviceRoleChanged,
    WatchdogWakeUpRequested,
    LensDistortionChanged,
    PropertyChanged(Property),
    WirelessDisconnect,
    WirelessReconnect,
    ButtonPress(Controller),
    ButtonUnpress(Controller),
    ButtonTouch(Controller),
    ButtonUntouch(Controller),
    DualAnalog_Press,
    DualAnalog_Unpress,
    DualAnalog_Touch,
    DualAnalog_Untouch,
    DualAnalog_Move,
    DualAnalog_ModeSwitch1,
    DualAnalog_ModeSwitch2,
    DualAnalog_Cancel,
    MouseMove(Mouse),
    MouseButtonDown(Mouse),
    MouseButtonUp(Mouse),
    FocusEnter(Overlay),
    FocusLeave(Overlay),
    ScrollDiscrete(Scroll),
    TouchPadMove(TouchPadMove),
    OverlayFocusChanged(Overlay),
    ReloadOverlays,
    ScrollSmooth(Scroll),
    InputFocusCaptured(Process),
    InputFocusReleased(Process),
    SceneFocusLost(Process),
    SceneFocusGained(Process),
    SceneApplicationChanged(Process),
    SceneFocusChanged(Process),
    InputFocusChanged(Process),
    SceneApplicationSecondaryRenderingStarted(Process),
    SceneApplicationUsingWrongGraphicsAdapter,
    ActionBindingReloaded,
    HideRenderModels,
    ShowRenderModels,
    ConsoleOpened,
    ConsoleClosed,
    OverlayShown,
    OverlayHidden,
    DashboardActivated,
    DashboardDeactivated,
    DashboardRequested,
    ResetDashboard,
    RenderToast,
    ImageLoaded,
    ShowKeyboard,
    HideKeyboard,
    OverlayGamepadFocusGained,
    OverlayGamepadFocusLost,
    OverlaySharedTextureChanged,
    ScreenshotTriggered,
    ImageFailed,
    DashboardOverlayCreated,
    SwitchGamepadFocus,
    RequestScreenshot,
    ScreenshotTaken,
    ScreenshotFailed,
    SubmitScreenshotToDashboard,
    ScreenshotProgressToDashboard,
    PrimaryDashboardDeviceChanged,
    RoomViewShown,
    RoomViewHidden,
    ShowUI,
    ShowDevTools,
    Notification_Shown,
    Notification_Hidden,
    Notification_BeginInteraction,
    Notification_Destroyed,
    Quit(Process),
    ProcessQuit(Process),
    QuitAborted_UserPrompt(Process),
    QuitAcknowledged(Process),
    DriverRequestedQuit,
    RestartRequested,
    ChaperoneDataHasChanged,
    ChaperoneUniverseHasChanged,
    ChaperoneTempDataHasChanged,
    ChaperoneSettingsHaveChanged,
    SeatedZeroPoseReset,
    ChaperoneFlushCache,
    ChaperoneRoomSetupStarting,
    ChaperoneRoomSetupFinished,
    AudioSettingsHaveChanged,
    BackgroundSettingHasChanged,
    CameraSettingsHaveChanged,
    ReprojectionSettingHasChanged,
    ModelSkinSettingsHaveChanged,
    EnvironmentSettingsHaveChanged,
    EnableHomeAppSettingsHaveChanged,
    PowerSettingsHaveChanged,
    SteamVRSectionSettingChanged,
    LighthouseSectionSettingChanged,
    NullSectionSettingChanged,
    UserInterfaceSectionSettingChanged,
    NotificationsSectionSettingChanged,
    KeyboardSectionSettingChanged,
    PerfSectionSettingChanged,
    DashboardSectionSettingChanged,
    WebInterfaceSectionSettingChanged,
    TrackersSectionSettingChanged,
    LastKnownSectionSettingChanged,
    DismissedWarningsSectionSettingChanged,
    StatusUpdate,
    WebInterface_InstallDriverCompleted,
    MCImageUpdated,
    FirmwareUpdateStarted,
    FirmwareUpdateFinished,
    KeyboardClosed,
    KeyboardCharInput(Keyboard),
    KeyboardDone,
    ApplicationTransitionStarted,
    ApplicationTransitionAborted,
    ApplicationTransitionNewAppStarted,
    ApplicationListUpdated,
    ApplicationMimeTypeLoad,
    ApplicationTransitionNewAppLaunchComplete,
    ProcessConnected,
    ProcessDisconnected,
    Compositor_MirrorWindowShown,
    Compositor_MirrorWindowHidden,
    Compositor_ChaperoneBoundsShown,
    Compositor_ChaperoneBoundsHidden,
    Compositor_DisplayDisconnected,
    Compositor_DisplayReconnected,
    Compositor_HDCPError,
    Compositor_ApplicationNotResponding,
    Compositor_ApplicationResumed,
    Compositor_OutOfVideoMemory,
    TrackedCamera_StartVideoStream,
    TrackedCamera_StopVideoStream,
    TrackedCamera_PauseVideoStream,
    TrackedCamera_ResumeVideoStream,
    TrackedCamera_EditingSurface,
    PerformanceTest_EnableCapture,
    PerformanceTest_DisableCapture,
    PerformanceTest_FidelityLevel,
    MessageOverlay_Closed,
    MessageOverlayCloseRequested,
    Input_HapticVibration,
    Input_BindingLoadFailed,
    Input_BindingLoadSuccessful,
    Input_ActionManifestReloaded,
    Input_ActionManifestLoadFailed,
    Input_ProgressUpdate,
    Input_TrackerActivated,
    Input_BindingsUpdated,
    SpatialAnchors_PoseUpdated,
    SpatialAnchors_DescriptorUpdated,
    SpatialAnchors_RequestPoseUpdate,
    SpatialAnchors_RequestDescriptorUpdate,
    SystemReport_Started,
    VendorSpecific(EVREventType),
    Unknown(EVREventType),
}

Variants

TrackedDeviceActivated

TrackedDeviceDeactivated

TrackedDeviceUpdated

TrackedDeviceUserInteractionStarted

TrackedDeviceUserInteractionEnded

IpdChanged

EnterStandbyMode

LeaveStandbyMode

TrackedDeviceRoleChanged

WatchdogWakeUpRequested

LensDistortionChanged

PropertyChanged(Property)

WirelessDisconnect

WirelessReconnect

ButtonPress(Controller)

ButtonUnpress(Controller)

ButtonTouch(Controller)

ButtonUntouch(Controller)

DualAnalog_Press

DualAnalog_Unpress

DualAnalog_Touch

DualAnalog_Untouch

DualAnalog_Move

DualAnalog_ModeSwitch1

DualAnalog_ModeSwitch2

DualAnalog_Cancel

MouseMove(Mouse)

MouseButtonDown(Mouse)

MouseButtonUp(Mouse)

FocusEnter(Overlay)

FocusLeave(Overlay)

ScrollDiscrete(Scroll)

TouchPadMove(TouchPadMove)

OverlayFocusChanged(Overlay)

global event

ReloadOverlays

ScrollSmooth(Scroll)

InputFocusCaptured(Process)

Deprecated

InputFocusReleased(Process)

Deprecated

SceneFocusLost(Process)

SceneFocusGained(Process)

SceneApplicationChanged(Process)

The app actually drawing the scene changed (usually to or from the compositor)

SceneFocusChanged(Process)

New app got access to draw the scene

InputFocusChanged(Process)

SceneApplicationSecondaryRenderingStarted(Process)

SceneApplicationUsingWrongGraphicsAdapter

ActionBindingReloaded

HideRenderModels

Sent to the scene application to request hiding render models temporarily

ShowRenderModels

Sent to the scene application to request restoring render model visibility

ConsoleOpened

ConsoleClosed

OverlayShown

OverlayHidden

DashboardActivated

DashboardDeactivated

DashboardRequested

Sent to the overlay manager - data is overlay

ResetDashboard

Send to the overlay manager

RenderToast

Send to the dashboard to render a toast - data is the notification ID

ImageLoaded

Sent to overlays when a SetOverlayRaw or SetOverlayFromFile call finishes loading

ShowKeyboard

Sent to keyboard renderer in the dashboard to invoke it

HideKeyboard

Sent to keyboard renderer in the dashboard to hide it

OverlayGamepadFocusGained

Sent to an overlay when IVROverlay::SetFocusOverlay is called on it

OverlayGamepadFocusLost

Send to an overlay when it previously had focus and IVROverlay::SetFocusOverlay is called on something else

OverlaySharedTextureChanged

ScreenshotTriggered

Screenshot button combo was pressed, Dashboard should request a screenshot

ImageFailed

Sent to overlays when a SetOverlayRaw or SetOverlayfromFail fails to load

DashboardOverlayCreated

SwitchGamepadFocus

RequestScreenshot

Sent by vrclient application to compositor to take a screenshot

ScreenshotTaken

Sent by compositor to the application that the screenshot has been taken

ScreenshotFailed

Sent by compositor to the application that the screenshot failed to be taken

SubmitScreenshotToDashboard

Sent by compositor to the dashboard that a completed screenshot was submitted

ScreenshotProgressToDashboard

Sent by compositor to the dashboard that a completed screenshot was submitted

PrimaryDashboardDeviceChanged

RoomViewShown

RoomViewHidden

ShowUI

ShowDevTools

Notification_Shown

Notification_Hidden

Notification_BeginInteraction

Notification_Destroyed

Quit(Process)

The application has been asked to quit

ProcessQuit(Process)

QuitAborted_UserPrompt(Process)

QuitAcknowledged(Process)

DriverRequestedQuit

The driver has requested that SteamVR shut down

RestartRequested

ChaperoneDataHasChanged

ChaperoneUniverseHasChanged

ChaperoneTempDataHasChanged

ChaperoneSettingsHaveChanged

SeatedZeroPoseReset

ChaperoneFlushCache

ChaperoneRoomSetupStarting

ChaperoneRoomSetupFinished

AudioSettingsHaveChanged

BackgroundSettingHasChanged

CameraSettingsHaveChanged

ReprojectionSettingHasChanged

ModelSkinSettingsHaveChanged

EnvironmentSettingsHaveChanged

EnableHomeAppSettingsHaveChanged

PowerSettingsHaveChanged

SteamVRSectionSettingChanged

LighthouseSectionSettingChanged

NullSectionSettingChanged

UserInterfaceSectionSettingChanged

NotificationsSectionSettingChanged

KeyboardSectionSettingChanged

PerfSectionSettingChanged

DashboardSectionSettingChanged

WebInterfaceSectionSettingChanged

TrackersSectionSettingChanged

LastKnownSectionSettingChanged

DismissedWarningsSectionSettingChanged

StatusUpdate

WebInterface_InstallDriverCompleted

MCImageUpdated

FirmwareUpdateStarted

FirmwareUpdateFinished

KeyboardClosed

KeyboardCharInput(Keyboard)

KeyboardDone

Sent when DONE button clicked on keyboard

ApplicationTransitionStarted

ApplicationTransitionAborted

ApplicationTransitionNewAppStarted

ApplicationListUpdated

ApplicationMimeTypeLoad

ApplicationTransitionNewAppLaunchComplete

ProcessConnected

ProcessDisconnected

Compositor_MirrorWindowShown

Compositor_MirrorWindowHidden

Compositor_ChaperoneBoundsShown

Compositor_ChaperoneBoundsHidden

Compositor_DisplayDisconnected

Compositor_DisplayReconnected

Compositor_HDCPError

Compositor_ApplicationNotResponding

Compositor_ApplicationResumed

Compositor_OutOfVideoMemory

TrackedCamera_StartVideoStream

TrackedCamera_StopVideoStream

TrackedCamera_PauseVideoStream

TrackedCamera_ResumeVideoStream

TrackedCamera_EditingSurface

PerformanceTest_EnableCapture

PerformanceTest_DisableCapture

PerformanceTest_FidelityLevel

MessageOverlay_Closed

MessageOverlayCloseRequested

Input_HapticVibration

Input_BindingLoadFailed

Input_BindingLoadSuccessful

Input_ActionManifestReloaded

Input_ActionManifestLoadFailed

Input_ProgressUpdate

Input_TrackerActivated

Input_BindingsUpdated

SpatialAnchors_PoseUpdated

SpatialAnchors_DescriptorUpdated

SpatialAnchors_RequestPoseUpdate

SpatialAnchors_RequestDescriptorUpdate

SystemReport_Started

VendorSpecific(EVREventType)

Unknown(EVREventType)