Crate openvr

Modules

button_id

compositor
The Compositor implements distortion, prediction, synchronization and other subtle issues that can be a challenge to get operating properly for a solid VR experience.

property

render_models

system
The System interface provides access to display configuration information, tracking data, controller state, events, and device properties. It is the main interface of OpenVR.

tracked_device_index

Structs

Compositor

Context
Entry points into OpenVR.

ControllerAxis

ControllerState

InitError

RenderModels

System

TrackedDevicePose

VkDevice_T

VkInstance_T

VkPhysicalDevice_T

VkQueue_T

Enums

ApplicationType

Eye

TrackedControllerRole

TrackedDeviceClass

TrackingResult

TrackingUniverseOrigin

Constants

MAX_TRACKED_DEVICE_COUNT

Functions

init⚠
Initialize OpenVR

Type Definitions

TrackedDeviceIndex

TrackedDevicePoses

TrackedDeviceProperty