Enum openvr::system::event::Event

pub enum Event {
    TrackedDeviceActivated,
    TrackedDeviceDeactivated,
    TrackedDeviceUpdated,
    TrackedDeviceUserInteractionStarted,
    TrackedDeviceUserInteractionEnded,
    IpdChanged,
    EnterStandbyMode,
    LeaveStandbyMode,
    TrackedDeviceRoleChanged,
    WatchdogWakeUpRequested,
    LensDistortionChanged,
    PropertyChanged(Property),
    ButtonPress(Controller),
    ButtonUnpress(Controller),
    ButtonTouch(Controller),
    ButtonUntouch(Controller),
    MouseMove(Mouse),
    MouseButtonDown(Mouse),
    MouseButtonUp(Mouse),
    FocusEnter(Overlay),
    FocusLeave(Overlay),
    Scroll(Scroll),
    TouchPadMove(TouchPadMove),
    OverlayFocusChanged(Overlay),
    InputFocusCaptured(Process),
    InputFocusReleased(Process),
    SceneFocusLost(Process),
    SceneFocusGained(Process),
    SceneApplicationChanged(Process),
    SceneFocusChanged(Process),
    InputFocusChanged(Process),
    SceneApplicationSecondaryRenderingStarted(Process),
    HideRenderModels,
    ShowRenderModels,
    OverlayShown,
    OverlayHidden,
    DashboardActivated,
    DashboardDeactivated,
    DashboardThumbSelected,
    DashboardRequested,
    ResetDashboard,
    RenderToast,
    ImageLoaded,
    ShowKeyboard,
    HideKeyboard,
    OverlayGamepadFocusGained,
    OverlayGamepadFocusLost,
    OverlaySharedTextureChanged,
    DashboardGuideButtonDown,
    DashboardGuideButtonUp,
    ScreenshotTriggered,
    ImageFailed,
    DashboardOverlayCreated,
    RequestScreenshot,
    ScreenshotTaken,
    ScreenshotFailed,
    SubmitScreenshotToDashboard,
    ScreenshotProgressToDashboard,
    PrimaryDashboardDeviceChanged,
    Notification_Shown,
    Notification_Hidden,
    Notification_BeginInteraction,
    Notification_Destroyed,
    Quit(Process),
    ProcessQuit(Process),
    QuitAborted_UserPrompt(Process),
    QuitAcknowledged(Process),
    DriverRequestedQuit,
    ChaperoneDataHasChanged,
    ChaperoneUniverseHasChanged,
    ChaperoneTempDataHasChanged,
    ChaperoneSettingsHaveChanged,
    SeatedZeroPoseReset,
    AudioSettingsHaveChanged,
    BackgroundSettingHasChanged,
    CameraSettingsHaveChanged,
    ReprojectionSettingHasChanged,
    ModelSkinSettingsHaveChanged,
    EnvironmentSettingsHaveChanged,
    PowerSettingsHaveChanged,
    StatusUpdate,
    MCImageUpdated,
    FirmwareUpdateStarted,
    FirmwareUpdateFinished,
    KeyboardClosed,
    KeyboardCharInput(Keyboard),
    KeyboardDone,
    ApplicationTransitionStarted,
    ApplicationTransitionAborted,
    ApplicationTransitionNewAppStarted,
    ApplicationListUpdated,
    ApplicationMimeTypeLoad,
    ApplicationTransitionNewAppLaunchComplete,
    ProcessConnected,
    ProcessDisconnected,
    Compositor_MirrorWindowShown,
    Compositor_MirrorWindowHidden,
    Compositor_ChaperoneBoundsShown,
    Compositor_ChaperoneBoundsHidden,
    TrackedCamera_StartVideoStream,
    TrackedCamera_StopVideoStream,
    TrackedCamera_PauseVideoStream,
    TrackedCamera_ResumeVideoStream,
    TrackedCamera_EditingSurface,
    PerformanceTest_EnableCapture,
    PerformanceTest_DisableCapture,
    PerformanceTest_FidelityLevel,
    MessageOverlay_Closed,
    VendorSpecific(u32),
    Unknown(u32),
}

Variants
TrackedDeviceActivatedTrackedDeviceDeactivatedTrackedDeviceUpdatedTrackedDeviceUserInteractionStartedTrackedDeviceUserInteractionEndedIpdChangedEnterStandbyModeLeaveStandbyModeTrackedDeviceRoleChangedWatchdogWakeUpRequestedLensDistortionChangedPropertyChanged(Property)ButtonPress(Controller)ButtonUnpress(Controller)ButtonTouch(Controller)ButtonUntouch(Controller)MouseMove(Mouse)MouseButtonDown(Mouse)MouseButtonUp(Mouse)FocusEnter(Overlay)FocusLeave(Overlay)Scroll(Scroll)TouchPadMove(TouchPadMove)OverlayFocusChanged(Overlay)

global event

InputFocusCaptured(Process)

Deprecated

InputFocusReleased(Process)

Deprecated

SceneFocusLost(Process)SceneFocusGained(Process)SceneApplicationChanged(Process)

The app actually drawing the scene changed (usually to or from the compositor)

SceneFocusChanged(Process)

New app got access to draw the scene

InputFocusChanged(Process)SceneApplicationSecondaryRenderingStarted(Process)HideRenderModels

Sent to the scene application to request hiding render models temporarily

ShowRenderModels

Sent to the scene application to request restoring render model visibility

OverlayShownOverlayHiddenDashboardActivatedDashboardDeactivatedDashboardThumbSelected

Sent to the overlay manager - data is overlay

DashboardRequested

Sent to the overlay manager - data is overlay

ResetDashboard

Send to the overlay manager

RenderToast

Send to the dashboard to render a toast - data is the notification ID

ImageLoaded

Sent to overlays when a SetOverlayRaw or SetOverlayFromFile call finishes loading

ShowKeyboard

Sent to keyboard renderer in the dashboard to invoke it

HideKeyboard

Sent to keyboard renderer in the dashboard to hide it

OverlayGamepadFocusGained

Sent to an overlay when IVROverlay::SetFocusOverlay is called on it

OverlayGamepadFocusLost

Send to an overlay when it previously had focus and IVROverlay::SetFocusOverlay is called on something else

OverlaySharedTextureChangedDashboardGuideButtonDownDashboardGuideButtonUpScreenshotTriggered

Screenshot button combo was pressed, Dashboard should request a screenshot

ImageFailed

Sent to overlays when a SetOverlayRaw or SetOverlayfromFail fails to load

DashboardOverlayCreatedRequestScreenshot

Sent by vrclient application to compositor to take a screenshot

ScreenshotTaken

Sent by compositor to the application that the screenshot has been taken

ScreenshotFailed

Sent by compositor to the application that the screenshot failed to be taken

SubmitScreenshotToDashboard

Sent by compositor to the dashboard that a completed screenshot was submitted

ScreenshotProgressToDashboard

Sent by compositor to the dashboard that a completed screenshot was submitted

PrimaryDashboardDeviceChangedNotification_ShownNotification_HiddenNotification_BeginInteractionNotification_DestroyedQuit(Process)

The application has been asked to quit

ProcessQuit(Process)QuitAborted_UserPrompt(Process)QuitAcknowledged(Process)DriverRequestedQuit

The driver has requested that SteamVR shut down

ChaperoneDataHasChangedChaperoneUniverseHasChangedChaperoneTempDataHasChangedChaperoneSettingsHaveChangedSeatedZeroPoseResetAudioSettingsHaveChangedBackgroundSettingHasChangedCameraSettingsHaveChangedReprojectionSettingHasChangedModelSkinSettingsHaveChangedEnvironmentSettingsHaveChangedPowerSettingsHaveChangedStatusUpdateMCImageUpdatedFirmwareUpdateStartedFirmwareUpdateFinishedKeyboardClosedKeyboardCharInput(Keyboard)KeyboardDone

Sent when DONE button clicked on keyboard

ApplicationTransitionStartedApplicationTransitionAbortedApplicationTransitionNewAppStartedApplicationListUpdatedApplicationMimeTypeLoadApplicationTransitionNewAppLaunchCompleteProcessConnectedProcessDisconnectedCompositor_MirrorWindowShownCompositor_MirrorWindowHiddenCompositor_ChaperoneBoundsShownCompositor_ChaperoneBoundsHiddenTrackedCamera_StartVideoStreamTrackedCamera_StopVideoStreamTrackedCamera_PauseVideoStreamTrackedCamera_ResumeVideoStreamTrackedCamera_EditingSurfacePerformanceTest_EnableCapturePerformanceTest_DisableCapturePerformanceTest_FidelityLevelMessageOverlay_ClosedVendorSpecific(u32)Unknown(u32)